% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{evaluate_glreg}
\alias{evaluate_glreg}
\title{Calculate metrics for a single glreg result object}
\usage{
evaluate_glreg(glreg_result, L_true)
}
\arguments{
\item{glreg_result}{A glreg object (output from glreg())}

\item{L_true}{True Laplacian matrix}
}
\value{
A graph_evaluation object
}
\description{
Convenience function to evaluate a glreg object directly
}
