% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_generation.R
\name{plot.glreg_signals}
\alias{plot.glreg_signals}
\title{Plot method for glreg_signals objects}
\usage{
\method{plot}{glreg_signals}(x, ...)
}
\arguments{
\item{x}{A glreg_signals object}

\item{...}{Additional arguments (not used)}
}
\description{
Plot method for glreg_signals objects
}
