% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{plot.method_comparison}
\alias{plot.method_comparison}
\title{Plot comparison of methods}
\usage{
\method{plot}{method_comparison}(x, ...)
}
\arguments{
\item{x}{A method_comparison object}

\item{...}{Additional arguments (not used)}
}
\description{
Plot comparison of methods
}
