% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glreg.R
\name{print.glreg}
\alias{print.glreg}
\title{Print method for glreg objects}
\usage{
\method{print}{glreg}(x, ...)
}
\arguments{
\item{x}{A glreg object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for glreg objects
}
