% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_generation.R
\name{print.glreg_signals}
\alias{print.glreg_signals}
\title{Print method for glreg_signals objects}
\usage{
\method{print}{glreg_signals}(x, ...)
}
\arguments{
\item{x}{A glreg_signals object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for glreg_signals objects
}
