% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{print.graph_evaluation}
\alias{print.graph_evaluation}
\title{Print method for graph_evaluation objects}
\usage{
\method{print}{graph_evaluation}(x, ...)
}
\arguments{
\item{x}{A graph_evaluation object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for graph_evaluation objects
}
