% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{print.method_comparison}
\alias{print.method_comparison}
\title{Print method for method_comparison objects}
\usage{
\method{print}{method_comparison}(x, ...)
}
\arguments{
\item{x}{A method_comparison object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for method_comparison objects
}
