% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glreg.R
\name{summary.glreg}
\alias{summary.glreg}
\title{Summary method for glreg objects}
\usage{
\method{summary}{glreg}(object, ...)
}
\arguments{
\item{object}{A glreg object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for glreg objects
}
